split model service
